clustermesh: rename internal package to common
The internal package has a special meaning in go, preventing it from
being imported outside of the clustermesh package. Yet, this is possibly
problematic as other packages cannot for instance depend on the types
therein defined.
