+++ pkg/clustermesh/clustermesh.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/cilium/cilium/api/v1/models"
 	"github.com/cilium/cilium/pkg/allocator"
-	"github.com/cilium/cilium/pkg/clustermesh/internal"
+	"github.com/cilium/cilium/pkg/clustermesh/common"
 	"github.com/cilium/cilium/pkg/clustermesh/types"
 	cmtypes "github.com/cilium/cilium/pkg/clustermesh/types"
 	"github.com/cilium/cilium/pkg/hive"
@@ -33,7 +33,7 @@ var log = logging.DefaultLogger.WithField(logfields.LogSubsys, subsystem)
 type Configuration struct {
 	cell.In
 
-	internal.Config
+	common.Config
 
 	// ClusterIDName is the id/name of the local cluster. This is used for logging and metrics
 	types.ClusterIDName
@@ -72,8 +72,8 @@ type Configuration struct {
 	// with remote clusters, to ensure their uniqueness.
 	ClusterIDsManager clusterIDsManager
 
-	Metrics         Metrics
-	InternalMetrics internal.Metrics
+	Metrics       Metrics
+	CommonMetrics common.Metrics
 }
 
 // RemoteIdentityWatcher is any type which provides identities that have been
@@ -100,8 +100,8 @@ type ClusterMesh struct {
 	// conf is the configuration, it is immutable after NewClusterMesh()
 	conf Configuration
 
-	// internal implements the common logic to connect to remote clusters.
-	internal internal.ClusterMesh
+	// common implements the common logic to connect to remote clusters.
+	common common.ClusterMesh
 
 	// globalServices is a list of all global services. The datastructure
 	// is protected by its own mutex inside the structure.
@@ -127,7 +127,7 @@ func NewClusterMesh(lifecycle hive.Lifecycle, c Configuration) *ClusterMesh {
 		),
 	}
 
-	cm.internal = internal.NewClusterMesh(internal.Configuration{
+	cm.common = common.NewClusterMesh(common.Configuration{
 		Config:                       c.Config,
 		ClusterIDName:                c.ClusterIDName,
 		ClusterSizeDependantInterval: c.ClusterSizeDependantInterval,
@@ -136,14 +136,14 @@ func NewClusterMesh(lifecycle hive.Lifecycle, c Configuration) *ClusterMesh {
 		NewRemoteCluster: cm.newRemoteCluster,
 
 		NodeName: nodeName,
-		Metrics:  c.InternalMetrics,
+		Metrics:  c.CommonMetrics,
 	})
 
-	lifecycle.Append(&cm.internal)
+	lifecycle.Append(&cm.common)
 	return cm
 }
 
-func (cm *ClusterMesh) newRemoteCluster(name string, status internal.StatusFunc) internal.RemoteCluster {
+func (cm *ClusterMesh) newRemoteCluster(name string, status common.StatusFunc) common.RemoteCluster {
 	rc := &remoteCluster{
 		name:    name,
 		mesh:    cm,
@@ -175,14 +175,14 @@ func (cm *ClusterMesh) newRemoteCluster(name string, status internal.StatusFunc)
 // NumReadyClusters returns the number of remote clusters to which a connection
 // has been established
 func (cm *ClusterMesh) NumReadyClusters() int {
-	return cm.internal.NumReadyClusters()
+	return cm.common.NumReadyClusters()
 }
 
 // ClustersSynced returns after all clusters were synchronized with the bpf
 // datapath.
 func (cm *ClusterMesh) ClustersSynced(ctx context.Context) error {
 	swgs := make([]*lock.StoppableWaitGroup, 0)
-	cm.internal.ForEachRemoteCluster(func(rci internal.RemoteCluster) error {
+	cm.common.ForEachRemoteCluster(func(rci common.RemoteCluster) error {
 		rc := rci.(*remoteCluster)
 		swgs = append(swgs, rc.swg)
 		return nil
@@ -204,7 +204,7 @@ func (cm *ClusterMesh) Status() (status *models.ClusterMeshStatus) {
 		NumGlobalServices: int64(cm.globalServices.size()),
 	}
 
-	cm.internal.ForEachRemoteCluster(func(rci internal.RemoteCluster) error {
+	cm.common.ForEachRemoteCluster(func(rci common.RemoteCluster) error {
 		rc := rci.(*remoteCluster)
 		status.Clusters = append(status.Clusters, rc.Status())
 		return nil
