+++ pkg/clustermesh/clustermesh_test.go
@@ -15,7 +15,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/cilium/cilium/pkg/clustermesh/internal"
+	"github.com/cilium/cilium/pkg/clustermesh/common"
 	"github.com/cilium/cilium/pkg/clustermesh/types"
 	cmtypes "github.com/cilium/cilium/pkg/clustermesh/types"
 	cmutils "github.com/cilium/cilium/pkg/clustermesh/utils"
@@ -140,7 +140,7 @@ func TestClusterMesh(t *testing.T) {
 
 	usedIDs := NewClusterMeshUsedIDs()
 	cm := NewClusterMesh(hivetest.Lifecycle(t), Configuration{
-		Config:                internal.Config{ClusterMeshConfig: dir},
+		Config:                common.Config{ClusterMeshConfig: dir},
 		ClusterIDName:         types.ClusterIDName{ClusterID: 255, ClusterName: "test2"},
 		NodeKeyCreator:        testNodeCreator,
 		NodeObserver:          &testObserver{},
@@ -148,7 +148,7 @@ func TestClusterMesh(t *testing.T) {
 		IPCache:               ipc,
 		ClusterIDsManager:     usedIDs,
 		Metrics:               newMetrics(),
-		InternalMetrics:       internal.MetricsProvider(subsystem)(),
+		CommonMetrics:         common.MetricsProvider(subsystem)(),
 	})
 	require.NotNil(t, cm, "Failed to initialize clustermesh")
