+++ pkg/clustermesh/common/metrics.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package internal
+package common
 
 import (
 	"github.com/cilium/cilium/pkg/metrics"
